[SPARK-5860][CORE] switch to BigInt instead of BigDecimal

+++ core/src/main/scala/org/apache/spark/rdd/JdbcRDD.scala
@@ -64,8 +64,8 @@ class JdbcRDD[T: ClassTag](
     // bounds are inclusive, hence the + 1 here and - 1 on end
     val length = 1 + upperBound - lowerBound
     (0 until numPartitions).map(i => {
-      val start = lowerBound + ((BigDecimal(i) * length) / numPartitions).toLong
-      val end = lowerBound + ((BigDecimal(i + 1) * length) / numPartitions).toLong - 1
+      val start = lowerBound + ((BigInt(i) * length) / numPartitions).toLong
+      val end = lowerBound + ((BigInt(i + 1) * length) / numPartitions).toLong - 1
       new JdbcPartition(i, start, end)
     }).toArray
   }
